Testing with quantisation.

+++ notebooks/02-ai-label-data-generation.py
@@ -35,7 +35,7 @@ from nyx.data_loaders import HumanEvaluatedDataLoader
 
 RANDOM_SEED = 42
 # PRECISION = torch.float32
-PRECISION_NAME = 'int8'
+PRECISION_NAME = 'float16'
 DEVICE = "cuda"
 LABELLER_MODEL = "Qwen/Qwen2-72B-Instruct-GPTQ-Int4"  # "Qwen/Qwen2-7B-Instruct"
 # "Qwen/Qwen2-0.5B-Instruct"
